backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
PyMuPDF==1.24.1
httpx==0.27.0
groq==1.2.0
